// components/common/pagePagination/pagePaginationStyle.scss.d.ts
export type Styles = {
  'arrowNexPrevious': string;
  'arrowNext': string;
  'arrowPrevious': string;
  'componentContainer': string;
  'itemsCount': string;
  'leftContainer': string;
  'lineHelpInfo': string;
  'paginationLine': string;
  'rightContainer': string;
  'select': string;
  'totalItemsContainer': string;
};

export type ClassNames = keyof Styles;

declare const styles: Styles;

export default styles;